修复了数据库读写失败的问题;修正了部分语言问题;删除了部分无用代码;解决了其他异常

+++ Model/Alexnet_ImageNet.py
@@ -40,7 +40,7 @@ def img_post_handler(adv, args):
     adv = (adv * std) + mean
     adv = adv * 255.0
 
-    adv = np.clip(adv, 0, 255).astype(np.uint8)
+    adv = np.clip(adv, 0, 255).astype(np.float32)
     return adv
